Imp Note on zero vector. Adding code to test manually constructed simple gridworld using ADDS and BDDs.
Note: From experiments, it turns out the compute PRE of a state that is a zero vector returns True vector. Thus, it is important to assigned zero vector, i.e, all bits are zero to the initial state or ensure that we do not assign zero vector to any state in the game.

Further, AbsSynthe and other AIG based tools in the SynthComp assign the initial states to be zero vector, i.e., the value of the latches corresponding to the initial state is all zero.